\name{aftsem-package}
\alias{aftsem-package}
\alias{aftsem}
\docType{package}
\title{
\packageTitle{aftsem}
}
\description{
\packageDescription{aftsem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aftsem}
\packageIndices{aftsem}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{aftsem}

Maintainer: \packageMaintainer{aftsem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
