% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsem_gehan.R
\name{gehan_estimation}
\alias{gehan_estimation}
\title{Gehan's Estimation for Survival Data}
\usage{
gehan_estimation(y, Z, delta, rsmat, m, init = FALSE)
}
\arguments{
\item{y}{A numeric vector of survival times.}

\item{Z}{A matrix of covariates}

\item{delta}{A numeric vector indicating censoring status}

\item{rsmat}{A resampling matrix}

\item{m}{Method for quantreg optimalization}

\item{init}{A logical value indicating whether to return the initial fit object
(default is `FALSE`). If `FALSE`, only the coefficients are returned.}
}
\value{
If `init = FALSE` and `change == 1`, returns a list with elements `INTERCEPT`,
  `RESID`, `ITERS`, `CONVERGED`, `BETA`. Otherwise, returns a matrix of resampled
  Gehan estimates.
}
\description{
This function performs Gehan's estimation of regression parameters proposed by Jin
}
\note{
This function is a slightly different version from the original by Zherzen Jin,
  part of the now not available `lss` program.
}
\author{
Zherzen Jin
}
