% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsem_summary.R
\name{print.summaryaftsem}
\alias{print.summaryaftsem}
\title{Print method for objects of class `summaryaftsem`}
\usage{
\method{print}{summaryaftsem}(x, ...)
}
\arguments{
\item{x}{An object of class `summaryaftsem`}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function prints object `summaryaftsem`
}
\description{
Print method for objects of class `summaryaftsem`
}
