% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsem_summary.R
\name{summary.aftsem}
\alias{summary.aftsem}
\title{Summary function for aftsem package}
\usage{
\method{summary}{aftsem}(object, ...)
}
\arguments{
\item{object}{An object of aftsemfit}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class `summaryaftsem` that contains summary information of the
fitted aftsem model.
}
\description{
Provides a summary of an aftsem model fit, including the model call,
residuals, initial and final coefficient estimates, method, convergence status,
number of iterations, number of observations, percent of censored observations,
and if available, the estimated covariance matrix of the coefficients, standard
deviations, z-values, and p-values for a Wald test.
}
