\name{aftsem-package}
\alias{aftsem-package}
\docType{package}
\title{
\packageTitle{aftsem}
}
\description{
\packageDescription{aftsem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aftsem}
\packageIndices{aftsem}

}
\author{
\packageAuthor{aftsem}

Maintainer: \packageMaintainer{aftsem}
}
\references{

 Buckley, J.; James, I. Linear Regression with Censored Data. Biometrika. 1979,  issn 00063444.

 Jin, Z., Lin, D.Y., Wei, L. J., and Ying, Z. (2003). Rank-based inference for the accelerated failure time models, Biometrika, 90, 341-353.
 
 Heller, G. Smoothed rank regression with censored data. Journal of the American Statistical Association. 2007
}

\keyword{ package }
\seealso{

 \code{\link[survival]{survfit}}
 \code{\link[aftgee]{aftgee}}
}

